# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface LegacyLoginRequest {
  phone: string;
  /**
     * @minLength 4
     * @maxLength 4
     */
  pin: string;
}

export interface LocalAuthUser {
  id: string;
  phone: string;
  name: string;
  /** @nullable */
  handle: string | null;
}

export interface LocalAuthResponse {
  user: LocalAuthUser;
  token: string;
}

export type CatalogEventRequestEventType = typeof CatalogEventRequestEventType[keyof typeof CatalogEventRequestEventType];


export const CatalogEventRequestEventType = {
  view: 'view',
  click: 'click',
} as const;

export interface CatalogEventRequest {
  businessSlug: string;
  eventType: CatalogEventRequestEventType;
  /**
     * @minLength 24
     * @maxLength 24
     */
  offeringKey?: string;
  /**
     * @minLength 16
     * @maxLength 128
     */
  visitorId: string;
}

export interface CatalogEventResponse {
  recorded: boolean;
}

export interface CatalogAnalyticsProduct {
  key: string;
  name: string;
  clicks: number;
}

export interface CatalogAnalytics {
  catalogVisitors: number;
  productClicks: number;
  products: CatalogAnalyticsProduct[];
}

export interface CatalogAnalyticsResponse {
  analytics: CatalogAnalytics;
}

export type BeginReplitLoginParams = {
returnTo?: string;
};