# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog/analytics:
    post:
      operationId: recordCatalogEvent
      tags: [Analytics]
      summary: Record an anonymous catalog view or product click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogEventRequest"
      responses:
        "202":
          description: Event accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogEventResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CatalogEventRequest:
      type: object
      required: [businessSlug, eventType, visitorId]
      properties:
        businessSlug:
          type: string
        eventType:
          type: string
          enum: [view, click]
        offeringKey:
          type: string
          minLength: 24
          maxLength: 24
        visitorId:
          type: string
          minLength: 16
          maxLength: 128
    CatalogEventResponse:
      type: object
      required: [recorded]
      properties:
        recorded:
          type: boolean
...