# lib/api-zod/src/generated/types/catalogEventRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogEventRequestEventType } from './catalogEventRequestEventType';

export interface CatalogEventRequest {
  businessSlug: string;
  eventType: CatalogEventRequestEventType;
  /**
     * @minLength 24
     * @maxLength 24
     */
  offeringKey?: string;
  /**
     * @minLength 16
     * @maxLength 128
     */
  visitorId: string;
}